requirements.txt
absl-py==0.9.0
asn1crypto==0.24.0
astor==0.8.1
attrs==19.3.0
Automat==0.6.0
backcall==0.1.0
bleach==3.1.5
blinker==1.4
certifi==2018.1.18
chardet==3.0.4
click==7.1.2
cloud-init==19.4
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
ec2-hibinit-agent==1.0.0
entrypoints==0.3
Flask==1.1.2
gast==0.3.3
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
hibagent==1.0.1
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
importlib-metadata==1.6.0
incremental==16.10.1
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
joblib==0.15.1
jsonpatch==1.16
jsonpointer==1.10
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.2.0
language-selector==0.1
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.6
netifaces==0.10.4
nltk==3.5
notebook==6.0.3
numpy==1.18.5
oauthlib==2.0.6
packaging==20.4
PAM==0.4.2
pandas==1.0.4
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
Pygments==2.6.1
pygobject==3.26.1
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyparsing==2.4.7
pyrsistent==0.16.0
pyserial==3.4
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2020.1
pyxdg==0.25
PyYAML==3.12
pyzmq==19.0.1
regex==2020.5.14
requests==2.18.4
requests-unixsocket==0.1.5
scikit-learn==0.23.1
scipy==1.4.1
SecretStorage==2.3.1
Send2Trash==1.5.0
service-identity==16.0.0
six==1.15.0
sklearn==0.0
ssh-import-id==5.7
systemd-python==234
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.0.4
tqdm==4.46.1
traitlets==4.3.3
Twisted==17.9.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
wcwidth==0.2.3
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
zope.interface==4.3.2
